modifications
I modified:
1- The Account Controller:
after login and register the role is sent with response so that in the frontend we can check the role to show the user his resources.
2- Dtos and Models: the UserdId is modeified to be string to be the same as in the user table.

3- I deleted The userType because we do not need it, we have roles and that is enough to know the type of the user.

4- TokenService: new claim is added which is the role of the user to be part of the token's data
we need it to be in the token, so that when the user request the endpoint we have to check his role if he is authorized to access that endpoint or not by extracting the role from the token which the backend received.