modifications
The dashboard for the business is ready.
placing bids in real time is fixed and now is functioning.

i modified the way to reach the auctions
the previous way was to send the state from results page to the auction details page, but that method does not refresh the state when it is updated.
so that i modified it to include the auction id in the url, rather than send the state, and get the auction data from the api once the user reach the auction details page.